+++ src/main/java/school/faang/user_service/dto/mentorshiprequest/RejectionDto.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.mentorshiprequest;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
