+++ src/main/java/school/faang/user_service/service/mentorshiprequest/MentorshipRequestService.java
@@ -1,17 +1,17 @@
-package school.faang.user_service.services;
+package school.faang.user_service.service.mentorshiprequest;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import school.faang.user_service.dto.MentorshipRequestDto;
-import school.faang.user_service.dto.RejectionDto;
-import school.faang.user_service.dto.RequestFilterDto;
+import school.faang.user_service.dto.mentorshiprequest.MentorshipRequestDto;
+import school.faang.user_service.dto.mentorshiprequest.RejectionDto;
+import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
 import school.faang.user_service.entity.RequestStatus;
-import school.faang.user_service.helper.filters.RequestFilter;
-import school.faang.user_service.helper.validator.MentorshipRequestValidator;
+import school.faang.user_service.helper.filters.mentorshiprequest.RequestFilter;
+import school.faang.user_service.helper.validator.mentorshiprequst.MentorshipRequestValidator;
 import school.faang.user_service.repository.mentorship.MentorshipRequestRepository;
-import school.faang.user_service.helper.mapper.MentorshipRequestMapper;
+import school.faang.user_service.helper.mapper.mentorshiprequest.MentorshipRequestMapper;
 
 import java.util.List;
 import java.util.Optional;
