+++ src/main/java/school/faang/user_service/service/user/UserService.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.services;
+package school.faang.user_service.service.user;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
