+++ src/test/java/school/faang/user_service/helper/filters/mentorshiprequest/RequestStatusFilterTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.helper.filters;
+package school.faang.user_service.helper.filters.mentorshiprequest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -6,10 +6,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
-import school.faang.user_service.dto.RequestFilterDto;
+import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
 import school.faang.user_service.entity.RequestStatus;
-import school.faang.user_service.helper.filters.RequestStatusFilter;
 
 import java.util.stream.Stream;
