+++ src/test/java/school/faang/user_service/helper/mapper/mentorshiprequest/MentorshipRequestMapperTest.java
@@ -1,9 +1,9 @@
-package school.faang.user_service.helper.mapper;
+package school.faang.user_service.helper.mapper.mentorshiprequest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
-import school.faang.user_service.dto.MentorshipRequestDto;
+import school.faang.user_service.dto.mentorshiprequest.MentorshipRequestDto;
 import school.faang.user_service.entity.MentorshipRequest;
 import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.User;
